refactor(lang_specs): extract extraction logic into `item_extractor`
