+++ tests/specs/item_extraction/item_extraction_spec.lua
@@ -1,7 +1,6 @@
 package.path = package.path .. ";" .. vim.fn.expand "<sfile>:p:h" .. "/tests"
 
 local test_utils = require "tests.utils"
-
 local LANGS_TO_TEST = require("codedocs").get_supported_langs()
 
 describe("Basic item extraction", function()
